+++ packages/react/src/components/NavTabs/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The NavTabs component represents the main navigation dock for an application on a mobile device, including up to 5 tabs or buttons for different sections. The component has a fixed position on the bottom of the screen.
 
+## 2025-09-17 `1.4.0`
+
+#### 🎉 New features
+
+- Added `ai-generative` appearance for `NavTabs.Item` component. ([#122](https://github.com/TiendaNube/nimbus-patterns/pull/122) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2025-06-04 `1.2.5`
 
 #### 💡 Others